refactor(app_partner): split event_action_card into focused files (#2146)
event_action_card.dart (466→332 lines):
- _EndedStatsRow + _StatCell → _event_action_card_stats.dart (part file)
- EventActionCardEmpty → event_action_card_empty.dart (standalone)
- Re-exported via event_action_card.dart for backward compatibility

Constraint: Test imports event_action_card.dart — re-export preserves API
Confidence: high
Scope-risk: narrow
